scripts/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "htd_scripts"
version = "0.0.1"
dependencies = [
    "bcrypt",
    "mysql-connector-python",
    "pandas",
    "sqlalchemy",
    "httpx",
    "pyaml",
]

[project.optional-dependencies]
dev = ["ruff", "nbstripout-fast"]
